refactor: change all TopAppBar to TopNavBar
