+++ src/components/TopNavBar/TopNavBar.style.ts
@@ -4,7 +4,7 @@ import { flexCenter } from '../../theme/functions';
 
 const DEFAULT_NAVBAR_HEIGHT = 62;
 
-const topAppBarWrapper = (dark: boolean) => (theme: Theme): SerializedStyles => css`
+const topNavBarWrapper = (dark: boolean) => (theme: Theme): SerializedStyles => css`
   ${flexCenter};
   position: relative;
   justify-content: space-between;
@@ -15,7 +15,7 @@ const topAppBarWrapper = (dark: boolean) => (theme: Theme): SerializedStyles => 
   box-shadow: ${theme.elevation['02']};
 `;
 
-const topAppBarSection = css`
+const topNavBarSection = css`
   ${flexCenter};
   flex-wrap: nowrap;
 `;
@@ -23,20 +23,20 @@ const topAppBarSection = css`
 const additionalToolsSection = (hasAdditionalTools: boolean) => (
   theme: Theme
 ): SerializedStyles => css`
-  ${topAppBarSection};
+  ${topNavBarSection};
   margin: 0 ${theme.spacing.md};
   flex-grow: ${hasAdditionalTools ? 1 : 0};
 `;
 
 const mainSection = (hasSearchBar: boolean) => css`
-  ${topAppBarSection};
+  ${topNavBarSection};
   flex-grow: ${hasSearchBar ? 1 : 0};
   justify-content: flex-start;
 `;
 
 export default {
-  topAppBarWrapper,
-  topAppBarSection,
+  topNavBarWrapper,
+  topNavBarSection,
   additionalToolsSection,
   mainSection,
 };
