+++ src/components/TopNavBar/components/SidebarMenuIcon/SidebarMenuIcon.style.ts
