+++ src/components/TopNavBar/components/SidebarMenuIcon/SidebarMenuIcon.tsx
