+++ src/components/TopNavBar/components/SidebarMenuIcon/index.ts
